cluster: split out nt_revision type
